backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy[asyncio]==2.0.34
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.3
gunicorn==21.2.0
fastapi-cors==0.0.6
dask[dataframe]==2024.9.0
psycopg2-binary==2.9.9
